SoundCloud
Was able to get music embedded from soundcloud using soundcloud API ; however, in order to get users to be able to like and comment on the songs from my web app I need to register my app with soundcloud and get a client id. Soundcloud is currently not accepting new registrations so functionality will be limited to listening to just listening to audio.